Prefer CollapsibleTitle over InventoryCollapsibleTitle
