+++ src/app/active-mode/LoadoutView.tsx
@@ -1,17 +1,12 @@
-import InventoryCollapsibleTitle from 'app/inventory/InventoryCollapsibleTitle';
+import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
 import { DimStore } from 'app/inventory/store-types';
 import LoadoutPopup from 'app/loadout/LoadoutPopup';
 import React from 'react';
 
 export default function LoadoutView({ store }: { store: DimStore }) {
   return (
-    <InventoryCollapsibleTitle
-      title={'Loadouts'}
-      sectionId={'Loadout'}
-      stores={[store]}
-      defaultCollapsed={true}
-    >
+    <CollapsibleTitle title={'Loadouts'} sectionId={'active-loadouts'} defaultCollapsed={true}>
       <LoadoutPopup dimStore={store} hideFarming={true} />
-    </InventoryCollapsibleTitle>
+    </CollapsibleTitle>
   );
 }
