+++ src/app/dim-ui/PageWithMenu.m.scss
@@ -54,6 +54,7 @@
   @include phone-portrait {
     position: static;
     margin-right: 0;
+    padding-right: 0;
     width: 100%;
     max-height: none;
   }
